Исправления по результатам ревью
